Small refactor of RoundCard and its components
